src/api-sap/requirements.txt
azure-functions
azure-storage-blob
azure-identity
azure-cosmosdb-table